Move tests closer to the code it covers
